requirements/global.txt
# Pululi
pulumi==1.8.1
pulumi_aws==1.17.0
boto3
# Code quality
pylint==2.4.4
autopep8==1.4.4
python-gnupg==0.4.5

--requirement lambda.txt
--requirement dynamic_providers.txt